Prevent user from cheating with DailyInput
We were not checking that the DailyInput being used for trading was the
same as the one we gave the algorithm. This was allowing the algorithm
to ignore the one provided and create a new one with $1 buy price or
$10,000,000 sell price.

We now check that the close price is the same as the one we provided to
prevent this spoofing.